Generify the method writeToCsv + enhanced the CswWriter
